requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.340
openai==1.3.5
langfuse==1.0.11
pydantic==2.5.2
